Fix NuppelVideoPlayer::IsNearEnd prototype following the libmythui-osd merge.
This fixes the reported compilation issue but IsNearEnd clearly needs a
little more work following the ProgramInfo refactor - there are
definitely some sign issues (refs #8527).

Refs #7650


git-svn-id: http://svn.mythtv.org/svn/trunk@24973 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/NuppelVideoPlayer.cpp
@@ -3160,7 +3160,7 @@ bool NuppelVideoPlayer::IsReallyNearEnd(void) const
  *  \param margin minimum number of frames we want before being near end,
  *                defaults to 2 seconds of video.
  */
-bool NuppelVideoPlayer::IsNearEnd(long long margin)
+bool NuppelVideoPlayer::IsNearEnd(int64_t margin)
 {
     uint64_t framesRead, framesLeft = 0;
